hook/requirements.txt
aliyun-python-sdk-core==2.15.1
aniso8601==9.0.1
apiutils==1.0.14
asgiref==3.8.1
attrs==24.3.0
blinker==1.8.2
certifi==2024.12.14
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
diff-match-patch==20230430
Django==5.0.11
django-constance==3.1.0
django-import-export==4.0.3
django-picklefield==3.2
django_restframework_apiview==1.4.3
djangorestframework==3.15.2
et-xmlfile==1.1.0
Flask==3.0.3
flask-core==2.9.0
Flask-Cors==5.0.0
flask-restx==1.3.0
greenlet==3.0.3
idna==3.10
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==0.10.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydal==20240428.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
tablib==3.5.0
toml==0.10.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.3.0
Werkzeug==3.0.6
redis==5.0.4
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
frozenlist==1.5.0
html2text==2024.2.26
multidict==6.1.0
propcache==0.2.1
py-cord==2.6.1
yarl==1.18.3
